Ajustando tipos da classe Account

+++ src/main/java/me/dio/domain/model/Account.java
@@ -1,5 +1,7 @@
 package me.dio.domain.model;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,6 +9,6 @@ public class Account {
   private Long id;
   private String number;
   private String agency;
-  private double balance;
-  private double limit;
+  private BigDecimal balance;
+  private BigDecimal limit;
 }
